Adding 'WebDriverWaitConditions.class' for containing common UI waits.
