+++ src/test/java/uk/gov/companieshouse/utils/ElementInteraction.java
@@ -1,23 +1,16 @@
 package uk.gov.companieshouse.utils;
 
-import static org.openqa.selenium.support.ui.ExpectedConditions.attributeToBeNotEmpty;
-import static org.openqa.selenium.support.ui.ExpectedConditions.visibilityOf;
-
-import java.time.Duration;
-
 import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.Keys;
-import org.openqa.selenium.TimeoutException;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.interactions.Actions;
-import org.openqa.selenium.support.ui.ExpectedConditions;
-import org.openqa.selenium.support.ui.WebDriverWait;
 
-public class ElementInteraction {
+public class ElementInteraction extends WebDriverWaitConditions {
 
     private final TestContext testContext;
 
     public ElementInteraction(TestContext testContext) {
+        super(testContext);
         this.testContext = testContext;
     }
 
@@ -50,37 +43,9 @@ public class ElementInteraction {
         actions.doubleClick(element).perform();
     }
 
-    public WebDriverWait getWebDriverWait(int secondsToWait) {
-        return new WebDriverWait(testContext.getWebDriver(), Duration.ofSeconds(secondsToWait));
-    }
 
     public void clearField(WebElement webElement) {
         webElement.sendKeys(Keys.chord(Keys.CONTROL,"a", Keys.DELETE));
-    }
-
-    public void waitForSpecificTextInElement(WebElement element, String text) {
-        try {
-            getWebDriverWait(10).until(ExpectedConditions.textToBePresentInElement(element, text));
-        } catch (TimeoutException exception) {
-            throw new TimeoutException("Expected value: " + text + " not found in element: " + element);
-        }
-    }
-
-    public void waitElementTextNotEmpty(WebElement element) {
-        try {
-            getWebDriverWait(10).until(attributeToBeNotEmpty(element, "value"));
-        } catch (TimeoutException exception) {
-            throw new TimeoutException("The following element was not populated with expected text: " + element);
-        }
-    }
-
-    public void waitUntilElementDisplayed(WebElement element) {
-        try {
-            getWebDriverWait(10).until(visibilityOf(element));
-        } catch (TimeoutException exception) {
-            throw new TimeoutException("The following element was not displayed as expected: " + element);
-        }
-
     }
 
 }
